Rename confirm-create-channel to confirm
The workflow is only responsible for waiting on user input to confirm or
reject.
