+++ packages/xstate-wallet/src/workflows/application.ts
@@ -15,7 +15,6 @@ import {
 
 import {MessagingServiceInterface, convertToChannelResult} from '../messaging';
 import {filter, map, tap, flatMap, first} from 'rxjs/operators';
-import * as CCC from './confirm-create-channel';
 import {
   createMockGuard,
   getDataAndInvoke2,
@@ -27,7 +26,7 @@ import {Store} from '../store';
 import {StateVariables} from '../store/types';
 import {ChannelStoreEntry} from '../store/channel-store-entry';
 import {bigNumberify} from 'ethers/utils';
-import {ConcludeChannel, CreateAndFund} from './';
+import {ConcludeChannel, CreateAndFund, Confirm as CCC} from './';
 
 import {
   PlayerStateUpdate,
